feat: move ExecutionRequest and ExecutionResponse structs to argus_models
