+++ crates/argus_models/src/lib.rs
@@ -1,3 +1,2 @@
-mod models;
-
-pub use models::monitor_match;
+pub mod monitor_match;
+pub mod js_executor;
