--- crates/argus_models/src/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod monitor_match;
